Fixing packaging issue preventing dependencies from being uploaded too (#2)
I discovered an issue where the compiled dependencies from layers and functions are not actually being uploaded.

The cause of this is because the handler is directing sam package to use the pre-build template (located at the template handler's "path" argument) rather than the post-build template (which will be in the pre-build template's .aws-sam directory). We don't actually need to specify the template path anyway, since it's only in special circumstances that this would ever be needed; sam package uses the right directory and template by default.